Only CLatLonCountryId needs float. JIRA: CXX-14376
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@106391 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/objects/seqfeat/SubSource.hpp
@@ -247,7 +247,7 @@ private:
 
     static string x_ParseDateRangeWithDelimiter(const string& orig_date, CTempString delim);
     static vector<string> x_GetDateTokens(const string& orig_date);
-    static CLatLonCountryId * x_CalculateLatLonId(float lat_value, float lon_value, string country, string province);
+    static CLatLonCountryId * x_CalculateLatLonId(double lat_value, double lon_value, string country, string province);
     static bool x_IsFixableIsoDate(const string& orig_date);
     static string x_RemoveIsoTime(const string& orig_date);
